BUG: Can only open tarfile in read mode if it exists

+++ squidle_downloader/download_images_tar.py
@@ -108,29 +108,32 @@ def download_images(
 
     os.makedirs(os.path.dirname(tar_fname), exist_ok=True)
 
-    try:
-        with tarfile.open(tar_fname, mode="r") as tar:
-            contents = tar.getnames()
-    except tarfile.ReadError:
-        t_wait = 5
-        if verbose >= 1:
-            print(
-                "{}Unable to open {}.\n{}File {} will be deleted in {} seconds..."
-                "".format(padding, tar_fname, padding, tar_fname, t_wait),
-            )
-            print("{}Deleting in".format(padding), end="", flush=True)
-            for i in range(t_wait // 1):
-                print(" {}...".format(t_wait - i), end="", flush=True)
-                time.sleep(1)
-            print(" Deleting!")
-            if t_wait % 1 > 0:
-                time.sleep(t_wait % 1)
-        else:
-            time.sleep(5)
-        os.remove(tar_fname)
-        if verbose >= 1:
-            print("{}Existing file {} deleted".format(padding, tar_fname), flush=True)
-        contents = {}
+    contents = {}
+    if os.path.exists(tar_fname):
+        try:
+            with tarfile.open(tar_fname, mode="r") as tar:
+                contents = tar.getnames()
+        except tarfile.ReadError:
+            t_wait = 5
+            if verbose >= 1:
+                print(
+                    "{}Unable to open {}.\n{}File {} will be deleted in {} seconds..."
+                    "".format(padding, tar_fname, padding, tar_fname, t_wait),
+                )
+                print("{}Deleting in".format(padding), end="", flush=True)
+                for i in range(t_wait // 1):
+                    print(" {}...".format(t_wait - i), end="", flush=True)
+                    time.sleep(1)
+                print(" Deleting!")
+                if t_wait % 1 > 0:
+                    time.sleep(t_wait % 1)
+            else:
+                time.sleep(5)
+            os.remove(tar_fname)
+            if verbose >= 1:
+                print(
+                    "{}Existing file {} deleted".format(padding, tar_fname), flush=True
+                )
 
     is_valid = np.zeros(len(df), dtype=bool)
     for i_row, (index, row) in enumerate(maybe_tqdm(df.iterrows())):
